common.inc added
You can now place a file with name "common.inc" in the calling directory and it will be sourced by the 00_common.sh automatically. So you can overwrite the setting-variables dynamically if you want. Also a new setting-variable calles "addrformat" was added to the 00_common.sh. This is normally fixed to "--mainnet" for shelley mainnet chain. But can be changed also to "--testnet-magic xxx" to produce testnet addresses with the scripts 02 and 03a instead of mainnet addresses.