Change parameter "skipUtxoWithPolicy" to "skipUtxoWithAsset"
updated the script 01_sendLovelaces.sh & 01_sendAssets.sh to be more precise in the "skip utxo" choice. its now possible to use "skipUtxoWithAsset: <policyID>(+<assetName)" in hex-format to exclude certain utxos from beeing involved in the transaction. so this command can be used to skip utxos with certain policies, or specific whole assets