README correction about StakeKeyRegistration
As it turns out, you don't need the stake.skey to register a stakeAddress on the chain. So you can register the stakeAddress of a HW Ledger on the chain also via normal CLI based payment keys. But if you wanna use a HW Wallet for the payment, it can only be used for its own stakeAddress.