Added encrypted Transaction Messages
Added the feature of encrypted transaction messages to the scripts:

- 01_sendLovelaces.sh
- 01_sendAssets.sh
- 01_claimRewards.sh
- 11a_mintAsset.sh
- 11b_burnAsset.sh

Encrypted transactions messages can be encrypted via the following modes: basic
This is done by adding the parameter "enc: basic" to the cli call.

In addition to the default passphrase 'cardano', the user can provide an own one. This is done by adding the passphrase via the new "pass:<passphrase>" parameter. Be aware, the passphrase starts right after the : char, so spaces are also included in the passphrase!