Update Catalyst_Registration_CLI_Tools.md
New updates on how to use Ledger HW-Wallets with the current Fund10 registration. The current cardano-app version is 5.0.1, the latest version 6.0.3 is not yet released. So we need to use cardano-hw-cli 1.12.0 for the registration and not the newer version cardano-hw-cli 1.13.0 ! 

Changes on how to do the registration have been updated in the MD file.