Update Catalyst_Registration_CLI_Tools.md
New updates on how to use Ledger HW-Wallets with the current Fund10 registration. The current cardano-app version is 5.0.1, the latest version 6.0.3 is not yet released. So we need to use cardano-hw-cli 1.12.0 for the registration and not the newer version cardano-hw-cli 1.13.0 ! 

Also the script 09a_catalystVote.sh has been updated to 09a_catalystVoteF10.sh to reflect the needed changes.

Changes on how to do the registration have been updated in the MD file. In case the cardano-app 6.0.3 will be release in time, the documentation/script will be updated again!