Adding new feature - Catalyst VotingPower query
09a_catalystVote.sh / 09a_vatalystVoteF10.sh:

Added a query feature so you can now verify your registered Voting-Power. This works on MainNet and on TestNet, because its a combined API at the moment.

The result is the currently registered total Voting-Power, and all the Delegators/Delegations that are done to that Voting-Key.

Queries can be done by providing the name of the Voting-Key file, by providing a bech encoded Voting-Key or by directly providing a hex encoded Voting-Key like:

- 09a_catalystVoteF10.sh query myvote-test

- 09a_catalystVoteF10.sh query cvote_vk1lca8dhe30dtmuk4yq3c9f20af0jk7ne9vhzhemheczjtpy3dq85qcdzn3e

- 09a_catalystVoteF10.sh query fe3a76df317b57be5aa4047054a9fd4be56f4f2565c57ceef9c0a4b0922d01e8