New features - Mnemonic wallet, Messages, Metadata
Generating Base/Staking-Keys with/from Mnemonics has been added to the script:

- 03a_genStakingPaymentAddr.sh

So you can now generate a set of Base(Payment)- and Staking-Keys derived from 24-Word Mnemonics. If Mnemonics are provided they are used, otherwise new ones are created. This makes it easier to use keys across the CLI and other wallets like Daedalus, Eternl, Typhoon, ...

---

Sending transaction messages including the option to encrypt the messages (CIP-0083) has been added to the following scripts:

- 03b_regStakingAddrCert.sh
- 05c_regStakepoolCert.sh
- 05d_poolWitness.sh
- 06_regDelegationCert.sh
- 07b_deregStakepoolCert.sh
- 08b_deregStakingAddrCert.sh

---

Attaching transaction metadata as *.json and *.cbor file has been added to the following scripts:

- 03b_regStakingAddrCert.sh
- 05c_regStakepoolCert.sh
- 05d_poolWitness.sh
- 06_regDelegationCert.sh
- 07b_deregStakepoolCert.sh
- 08b_deregStakingAddrCert.sh

So you can now add a comment or an encrypted message or metadata to every transaction that is possible via the StakepoolOperatorScripts.