Testnet - Added option to use own koiosApiToken in available
The scripts are using the publicly accessable koios Rest-API to gather information for example about a poolName, poolTicker, etc.
If you don't provide an own koiosAPI-Token, the scripts are automatically using the Public-Tier with koios request. On the other hand, if you have an own koiosAPI-Token, the scripts will use that and will also show you the Project-ID, the choosen Tier and the Expire-Date of your koiosAPI-Token.
You can set the koiosAPI-Token in the 00_common.sh scripts or via the common.inc like koiosApiToken="xxx"