Mainnet - More accurate fee calc for upcoming cardano-cli // use own koiosAPI-Token
Changes about fee calculation:

- more accurate dummy tx generation in all scripts with transactions for the upcoming correction in a new cardano-cli release (https://github.com/IntersectMBO/cardano-cli/pull/534)
- removal of the optional '--byron-witness-count' parameter

Use your own koiosAPI-Token if available:
The scripts are using the publicly accessable koios Rest-API to gather informations for example about a poolName, poolTicker, etc.
If you don't provide an own koiosAPI-Token, the scripts are automatically using the Public-Tier with koios request. On the other hand, if you have an own koiosAPI-Token, the scripts will use that and in "Light-Mode" they will also show you the Project-ID, the choosen Tier and the Expire-Date of your koiosAPI-Token.
You can set the koiosAPI-Token in the 00_common.sh scripts or via the common.inc like koiosApiToken="xxx"

Others changes:
- script 13b was missing some adahandle types, fixed