Update 2025-12 - ImportHelper Script for ConwayEra
The 0x_importHelper.sh script was updated to be compatible with latest cardano-cli release 10.12.0.0. The script was also lacking the requirement to set the correct stakeaddress-deposit-fee value while generating the xxx.deleg.cert file during the import process. Also some minor changes in the calling commands for cardano-cli.

Other changes:
- Minimum cardano-hw-cli version is not set to 1.19.0 to be in line with the new trezor message signing functions.